+++ example/lib/colors.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+const Color kcBackground = Color(0xFF16213E);
+const Color kcPrimary = Color(0xFF0F3460);
+const Color kcSecondary = Color(0xFF533483);
+const Color kcAccent = Color(0xFFE94560);