+++ lib/simple_sidebar_item.dart
@@ -5,9 +5,6 @@ class SimpleSidebarItem extends SimpleSidebarElement {
   /// The title of the item
   final String title;
 
-  /// The child that will be displayed when the item is selected
-  final Widget child;
-
   /// If the text should soft wrap or not
   final bool? wrapWord;
 
@@ -17,22 +14,32 @@ class SimpleSidebarItem extends SimpleSidebarElement {
   /// The custom tooltip message (default the title)
   final String? customTooltipMessage;
 
+  /// The widget that will be displayed in the front of the title
+  final Widget? leading;
+
+  /// The widget that will be displayed in the end of the title
+  final Widget? trailing;
+
   /// The icon that will be displayed in the front of the title
-  final IconData? iconFront;
+  final IconData? leadingIcon;
 
   /// The icon that will be displayed in the end of the title
-  final IconData? iconEnd;
+  final IconData? trailingIcon;
 
   /// Creates a new [SimpleSidebarItem]
-  SimpleSidebarItem(
-      {required this.title,
-      required this.child,
-      this.wrapWord,
-      this.textOverflow,
-      this.customTooltipMessage,
-      this.iconFront,
-      this.iconEnd}) {
-    assert(iconEnd != null || iconFront != null,
-        "You need to provide at least one icon");
+  SimpleSidebarItem({
+    required this.title,
+    this.wrapWord,
+    this.textOverflow,
+    this.customTooltipMessage,
+    this.leading,
+    this.trailing,
+    this.leadingIcon,
+    this.trailingIcon,
+  }) {
+    assert(!(trailing != null && trailingIcon != null),
+        "You can't use both trailing and trailingIcon");
+    assert(!(leading != null && leadingIcon != null),
+        "You can't use both leading and leadingIcon");
   }
 }