+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 ## 0.0.6
 * Removed:
   * Unwanted log (printed index every time tapped got called)
-* Fixed:
+* Fixed
   * always red background from the titleimage
   * selectedTextColor & unselectedTextColor not used
 * Added
@@ -37,3 +37,16 @@
 * Known bugs:
   * Not able to change the size of the titleImage
   * Not able to place elements on the bottom (footer)
+  
+## 0.0.7
+* Fixed:
+  * selectedTextColor & unselectedTextColor was still not used
+
+## 0.0.8
+* Improvement
+  * Reworked SimpleSidebarThemeData
+  * Renamed iconFront & iconEnd to leading, trailing and switched to Widget
+* Removed:
+  * Unused child in SimpleSidebarItem
+* Added:
+  * FooterItems (eg. Logout, Settings, etc.)